% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.r
\name{emc_record}
\alias{emc_record}
\title{Record an expression}
\usage{
emc_record(..., .label = NULL, .verbose = TRUE, .sep = "_",
  .cache_path = NULL)
}
\value{
A tibble that contains the result, error, execution times and the replay log.
}
\description{
Executes an expression recording all messages, warnings and errors along the way.
}
