% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{emc_start_cluster}
\alias{emc_start_cluster}
\title{Start a local cluster and register it with the multi-tasking backend.}
\usage{
emc_start_cluster(p = 3/4, envir = .GlobalEnv)
}
\arguments{
\item{p}{Ratio of the available logical cores to use. Rounds down to a minimum of 2.}

\item{envir}{Environment to save the cluster to. Defaults to the global environment.}
}
\description{
The cluster is registered with the global environment. Only one cluster can be running
at one time.
}
