% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{emc_stop_cluster}
\alias{emc_stop_cluster}
\title{Stop the local cluster}
\usage{
emc_stop_cluster(envir = .GlobalEnv)
}
\arguments{
\item{envir}{Environment the cluster was registered with}
}
\description{
Stop the local cluster
}
