
%%% $Id: print.Rd,v 1.5 2008/02/03 04:18:53 goswami Exp $

\name{print}

\alias{print.EMC}

\alias{print.EMCMaxTemper}

\alias{print.EMCPlaceTempers}

\title{The printing family of functions}

\description{
  The printing family of functions for this package.
}

\usage{
\method{print}{EMC}(x, \dots)
\method{print}{EMCMaxTemper}(x, \dots)
\method{print}{EMCPlaceTempers}(x, \dots)
}

\arguments{
  \item{x}{an object inheriting from class \code{EMC} (generated by
    functions \code{randomWalkMetropolis}, \code{MetropolisHastings},
    \code{parallelTempering} and \code{evolMonteCarlo}),
    \code{EMCMaxTemper} (generated by function \code{findMaxTemper}) or
    \code{EMCPlaceTempers} (generated by function \code{placeTempers}).}

  \item{\dots}{optional arguments passed to \code{print.default}; see
    its documentation.}
}

\author{Gopi Goswami \email{goswami@stat.harvard.edu}}

\seealso{\code{\link{randomWalkMetropolis}},
  \code{\link{MetropolisHastings}}, \code{\link{parallelTempering}},
  \code{\link{evolMonteCarlo}}, \code{\link{findMaxTemper}},
  \code{\link{placeTempers}}}

\keyword{print}
